requirements.txt
# Web framework
flask>=2.3,<4.0

# Google Gemini (GenAI SDK)
google-genai>=0.3,<1.0

# Environment variable support (optional but good practice)
python-dotenv>=1.0,<2.0
